Pull effectiveTimeFrame up into BaseDataPoint
