+++ java-schema-objects/src/main/java/org/openmhealth/schema/pojos/BodyHeight.java
@@ -20,10 +20,8 @@ import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.annotation.JsonRootName;
-import org.joda.time.DateTime;
 import org.openmhealth.schema.pojos.generic.DescriptiveStatistic;
 import org.openmhealth.schema.pojos.generic.LengthUnitValue;
-import org.openmhealth.schema.pojos.generic.TimeFrame;
 
 
 /**
@@ -36,9 +34,6 @@ public class BodyHeight extends BaseDataPoint {
     @JsonProperty(value = "body_height", required = true)
     private LengthUnitValue lengthUnitValue;
 
-    @JsonProperty(value = "effective_time_frame", required = false)
-    private TimeFrame effectiveTimeFrame;
-
     @JsonProperty(value = "descriptive_statistic", required = false)
     private DescriptiveStatistic descriptiveStatistic;
 
@@ -47,21 +42,10 @@ public class BodyHeight extends BaseDataPoint {
     public BodyHeight() {
     }
 
-    public BodyHeight(LengthUnitValue lengthUnitValue, TimeFrame effectiveTimeFrame) {
-        this.lengthUnitValue = lengthUnitValue;
-        this.effectiveTimeFrame = effectiveTimeFrame;
-    }
-
     @Override
     @JsonIgnore
     public String getSchemaName() {
         return SCHEMA_BODY_HEIGHT;
-    }
-
-    @Override
-    @JsonIgnore
-    public DateTime getTimeStamp() {
-        return effectiveTimeFrame.getTimestamp();
     }
 
     public DescriptiveStatistic getDescriptiveStatistic() {
@@ -78,13 +62,5 @@ public class BodyHeight extends BaseDataPoint {
 
     public void setLengthUnitValue(LengthUnitValue lengthUnitValue) {
         this.lengthUnitValue = lengthUnitValue;
-    }
-
-    public TimeFrame getEffectiveTimeFrame() {
-        return effectiveTimeFrame;
-    }
-
-    public void setEffectiveTimeFrame(TimeFrame effectiveTimeFrame) {
-        this.effectiveTimeFrame = effectiveTimeFrame;
     }
 }
