+++ java-schema-objects/src/main/java/org/openmhealth/schema/pojos/HeartRate.java
@@ -22,9 +22,7 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.annotation.JsonRootName;
 import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
 import com.fasterxml.jackson.databind.annotation.JsonSerialize;
-import org.joda.time.DateTime;
 import org.openmhealth.schema.pojos.generic.DescriptiveStatistic;
-import org.openmhealth.schema.pojos.generic.TimeFrame;
 import org.openmhealth.schema.pojos.serialize.LabeledEnumSerializer;
 import org.openmhealth.schema.pojos.serialize.TemporalRelationshipToPhysicalActivityDeserializer;
 
@@ -35,9 +33,6 @@ import org.openmhealth.schema.pojos.serialize.TemporalRelationshipToPhysicalActi
 @JsonInclude(JsonInclude.Include.NON_NULL)
 @JsonRootName(value = HeartRate.SCHEMA_HEART_RATE, namespace = DataPoint.NAMESPACE)
 public class HeartRate extends BaseDataPoint {
-
-    @JsonProperty(value = "effective_time_frame", required = false)
-    private TimeFrame effectiveTimeFrame;
 
     @JsonProperty(value = "descriptive_statistic", required = false)
     private DescriptiveStatistic descriptiveStatistic;
@@ -68,26 +63,12 @@ public class HeartRate extends BaseDataPoint {
         return SCHEMA_HEART_RATE;
     }
 
-    @Override
-    @JsonIgnore
-    public DateTime getTimeStamp() {
-        return effectiveTimeFrame.getTimestamp();
-    }
-
     public DescriptiveStatistic getDescriptiveStatistic() {
         return descriptiveStatistic;
     }
 
     public void setDescriptiveStatistic(DescriptiveStatistic descriptiveStatistic) {
         this.descriptiveStatistic = descriptiveStatistic;
-    }
-
-    public TimeFrame getEffectiveTimeFrame() {
-        return effectiveTimeFrame;
-    }
-
-    public void setEffectiveTimeFrame(TimeFrame effectiveTimeFrame) {
-        this.effectiveTimeFrame = effectiveTimeFrame;
     }
 
     public String getUserNotes() {
