+++ java-schema-objects/src/main/java/org/openmhealth/schema/pojos/SleepDuration.java
@@ -20,8 +20,6 @@ import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.annotation.JsonRootName;
-import org.joda.time.DateTime;
-import org.openmhealth.schema.pojos.generic.TimeFrame;
 
 
 /**
@@ -30,9 +28,6 @@ import org.openmhealth.schema.pojos.generic.TimeFrame;
 @JsonInclude(JsonInclude.Include.NON_NULL)
 @JsonRootName(value = SleepDuration.SCHEMA_SLEEP_DURATION, namespace = DataPoint.NAMESPACE)
 public class SleepDuration extends BaseDataPoint {
-
-    @JsonProperty(value = "effective_time_frame")
-    private TimeFrame effectiveTimeFrame;
 
     @JsonProperty(value = "sleep_duration")
     private SleepDurationUnitValue sleepDurationUnitValue;
@@ -49,20 +44,6 @@ public class SleepDuration extends BaseDataPoint {
     @JsonIgnore
     public String getSchemaName() {
         return SCHEMA_SLEEP_DURATION;
-    }
-
-    @Override
-    @JsonIgnore
-    public DateTime getTimeStamp() {
-        return effectiveTimeFrame.getTimestamp();
-    }
-
-    public TimeFrame getEffectiveTimeFrame() {
-        return effectiveTimeFrame;
-    }
-
-    public void setEffectiveTimeFrame(TimeFrame effectiveTimeFrame) {
-        this.effectiveTimeFrame = effectiveTimeFrame;
     }
 
     public SleepDurationUnitValue getSleepDurationUnitValue() {
